*Modified gHIamp source module in order to conform to g.tec's API example code, and possibly fix an amplifier-not-found bug. *Added SourceChGain and SourceChOffset parameters to the gHIamp source module in order to provide correct default values.
git-svn-id: http://www.bci2000.org/svn/trunk@4245 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/SignalSource/gHIampSource/gHIampADC.h
@@ -52,7 +52,7 @@ class gHIampADC : public BufferedADC
   typedef std::map< int, std::set< int > > ModeMap;
   ModeMap ParseModes( std::string modes ) const;
 
-  int mMasterIdx;
+  size_t mMasterIdx;
   gHIampDeviceContainer mDevices;
 
   // A little class for parsing SourceChList
